# backend/src/api/openapi-generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vibecode AI Planner API",
    "version": "1.0.0",
    "description": "API documentation for Vibecode AI Planner - project and ticket management with agent orchestration"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Agent API key for agent authentication"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/providers/{id}": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "Get a provider by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider details"
          }
        }
      },
      "patch": {
        "tags": [
          "Providers"
        ],
        "summary": "Update a provider",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Providers"
        ],
        "summary": "Delete a provider",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider deleted"
          }
        }
      }
    },
...
  },
...
}